rwcalrequirements.txt
Django==1.4
MySQL-python==1.2.3
URLObject==2.0.2
amqplib==1.0.2
anyjson==0.3.1
celery==2.5.3
django-celery==2.5.5
django-picklefield==0.2.1
gdata==2.0.17
google-api-python-client==1.0beta8
httplib2==0.7.4
importlib==1.0.2
kombu==2.1.8
oauth2==1.5.211
ordereddict==1.1
python-dateutil==1.5
python-gflags==2.0
pytz==2012c
wsgiref==0.1.2
